backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pandas>=2.2.0
openpyxl>=3.1.2
apscheduler>=3.10.4
pydantic-settings>=2.1.0
python-multipart>=0.0.9
pytest
httpx
